move aligners to separate folder and general clean up
- Moved Gotoh, Global, and WSB aligners to their own folder
- Moved Local aligner to own folder
- Added Gotoh and WSB examples
- Removed needless assignment of match_val, up_val, and left_val in
aligners
- Removed extra i and j assignment in Gotoh branch of select_aligner
function
