+++ goombay-rs/src/alignment/aligners/local/mod.rs
@@ -1,0 +1,3 @@
+pub mod base;
+
+pub use base::LocalAligner;
